Simplificando a pagina de login
